refactor: Adventure Domain package
